Add custom type for export format in export.go, and move cobra custom types to custom_types.go
